31 user notification integration

+++ backend-fastify/src/controllers/properties/delete-property.js
@@ -1,18 +1,28 @@
 import { ActivityType } from "../../enums/activity.js";
-import { NotificationType } from "../../enums/notifications.js";
+import { SocketNotificationType } from "../../enums/notifications.js";
 import { Property } from "../../models/property.js";
-import { createActivity } from "../../services/activity.js";
+import { User } from "../../models/user.js";
+import { addActivity } from "../../services/activity.js";
 import { activityPropertyDescription } from "../../utils/activity/index.js";
 import { authBearerToken } from "../../utils/requests.js";
 import { userIdToken } from "../../utils/users.js";
 import { sendTargetedNotification } from "../../websocket/index.js";
 import { unlinkImages } from "./image-property.js";
 
+/**
+ *
+ * @param {import("fastify").FastifyRequest} req
+ * @param {import("fastify").FastifyReply} res
+ */
 export const deleteProperty = async function (req, res) {
   const { id } = req.params;
   try {
     const token = authBearerToken(req);
     const user_id = userIdToken(token);
+    const user = await User.findOne({ user_id });
+    if (!user) {
+      return res.status(404).send({ message: "Error: User not found." });
+    }
     const property = await Property.findOneAndDelete({
       property_id: id,
       user_id,
@@ -25,17 +35,23 @@ export const deleteProperty = async function (req, res) {
     }
 
     // We Log User activity
-    const activity = await createActivity({
+    user.properties = user.properties.filter((i) => i !== id);
+    const activity = addActivity(user, {
       action: ActivityType.property.delete,
       description: activityPropertyDescription(
         ActivityType.property.delete,
         property
       ),
-      user_id,
       property_id: id,
     });
+    await user.save();
+
     if (activity) {
-      sendTargetedNotification(NotificationType.activity, activity, user_id);
+      sendTargetedNotification(
+        SocketNotificationType.activity,
+        activity,
+        user_id
+      );
     }
     return res.status(200).send({ data: { ...property.toObject() } });
   } catch (error) {
